+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/SelectableDataGrid.test.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
 import DataGrid from 'terra-clinical-data-grid';
+import classNames from 'classnames/bind';
 
 import ContentCellLayout from './ContentCellLayout';
+import styles from './ClinicalDataGridCommon.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 class SelectableDataGrid extends React.Component {
   constructor(props) {
@@ -89,12 +93,11 @@ class SelectableDataGrid extends React.Component {
     };
   }
 
-  /* eslint-disable react/forbid-dom-props */
   render() {
     const { columns } = this.state;
 
     return (
-      <div id="selectable-data-grid" style={{ height: '100%', width: '100%' }}>
+      <div id="selectable-data-grid" className={cx('content-wrapper')}>
         <DataGrid
           id="selections-example"
           pinnedColumns={[
@@ -158,7 +161,6 @@ class SelectableDataGrid extends React.Component {
       </div>
     );
   }
-  /* eslint-enable react/forbid-dom-props */
 }
 
 export default SelectableDataGrid;